refactor: #194 updated map structure
